% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{attach_apply_nodes}
\alias{attach_apply_nodes}
\title{Attach Apply nodes to a parent node.}
\usage{
attach_apply_nodes(
  all_var_details_rows,
  parent_node,
  db_name,
  custom_function_names,
  table_names
)
}
\arguments{
\item{all_var_details_rows}{Variable details rows associated with a variable.}

\item{parent_node}{An XML node.}

\item{db_name}{Database name.}

\item{custom_function_names}{vector of strings. Holds the names of functions
parsed from custom function files.}

\item{table_names}{vector of strings. Holds the names of tables included
in this model}
}
\value{
Updated parent node.
}
\description{
Attach Apply nodes to a parent node.
}
