% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{build_derived_field_node}
\alias{build_derived_field_node}
\title{Build DerivedField node.}
\usage{
build_derived_field_node(
  vars_sheet,
  var_details_sheet,
  var_name,
  db_name,
  custom_function_names,
  table_names
)
}
\arguments{
\item{vars_sheet}{Variables sheet data frame.}

\item{var_details_sheet}{Variable details sheet data frame.}

\item{var_name}{Variable name.}

\item{db_name}{Database name.}

\item{custom_function_names}{vector of strings. Holds the names of functions
parsed from custom function files.}

\item{table_names}{vector of strings. The list of tables part of this PMML}
}
\value{
DerivedField node.
}
\description{
Build DerivedField node.
}
