% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{build_node_for_rec_to}
\alias{build_node_for_rec_to}
\title{Constructs the PMML node to use for a value in a recEnd column}
\usage{
build_node_for_rec_to(var_details_row, db_name)
}
\arguments{
\item{var_details_row}{The row in a variable details sheet whose recEnd
column we will parse}

\item{db_name}{The database of the start variable}
}
\value{
list Contains the PMML node which can be either a Constant or
FieldRef node
}
\description{
Constructs the PMML node to use for a value in a recEnd column
}
