% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{build_trans_dict}
\alias{build_trans_dict}
\title{Build a TransformationDictionary node.}
\usage{
build_trans_dict(
  vars_sheet,
  var_details_sheet,
  var_names,
  db_name,
  custom_function_names,
  table_names
)
}
\arguments{
\item{vars_sheet}{Variable sheet data frame.}

\item{var_details_sheet}{Variable details sheet data frame.}

\item{var_names}{Vector of variable names.}

\item{db_name}{Database name.}

\item{custom_function_names}{vector of strings. Holds the names of functions
parsed from custom function files.}

\item{table_names}{vector of strings. The list of tables part of this PMML}
}
\value{
TransformationDictionary node.
}
\description{
Build a TransformationDictionary node.
}
