% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-to-pmml-switch.R
\name{convert_switch_statement_to_pmml}
\alias{convert_switch_statement_to_pmml}
\title{Converts a switch statement into its corresponding PMML}
\usage{
convert_switch_statement_to_pmml(
  expr,
  tokens,
  scope_variables,
  get_pmml_str_for_expr,
  get_pmml_str_for_token
)
}
\arguments{
\item{expr}{A data.frame with the expr node that holds the switch statement}

\item{tokens}{A data.frame containing the master list of tokens}

\item{scope_variables}{Pass through variable}

\item{get_pmml_str_for_expr}{Pass through variable}

\item{get_pmml_str_for_token}{Pass through variable}
}
\description{
Converts a switch statement into its corresponding PMML
}
