% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-model-export-to-pmml.R
\name{create_baseline_hazards_nodes}
\alias{create_baseline_hazards_nodes}
\title{Create Baseline Hazards Nodes}
\usage{
create_baseline_hazards_nodes(time, baseline_hazard, max_time, PMML)
}
\arguments{
\item{time}{value of the corresponding time}

\item{baseline_hazard}{value of the baseline_hazard}

\item{max_time}{largest time recTo value in variable_details}

\item{PMML}{the pmml that is appended to}
}
\value{
returns PMML with attached nodes correlating to the baseline_hazards
}
\description{
Creates nodes correlating to the baseline_hazards parameters
}
