% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-model-export-to-pmml.R
\name{create_beta_coefficient_nodes}
\alias{create_beta_coefficient_nodes}
\title{Create Beta Coefficient Nodes}
\usage{
create_beta_coefficient_nodes(variable, coefficient, type, mining_fields, PMML)
}
\arguments{
\item{variable}{name of the original variable}

\item{coefficient}{the coefficient used for the transformation}

\item{type}{type of the variable}

\item{PMML}{the pmml that is appended to}
}
\value{
returns PMML with attached nodes correlating to the beta_coefficient
}
\description{
Creates nodes correlating to the beta_coefficient parameters
}
