% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-model-export-to-pmml.R
\name{create_dummy_nodes}
\alias{create_dummy_nodes}
\title{Create Dummy Nodes}
\usage{
create_dummy_nodes(orig_variable, cat_value, dummy_variable, PMML)
}
\arguments{
\item{orig_variable}{a vector representing the original variable name}

\item{cat_value}{a vector representing the original variable cat value}

\item{dummy_variable}{a vector representing the new dummy variable name}

\item{PMML}{to append to}
}
\value{
PMML containing the newly added DerivedFields
}
\description{
Creates dummy nodes given passed parameters. The passed params must all be of same length
}
