% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-model-export-to-pmml.R
\name{create_interaction_nodes}
\alias{create_interaction_nodes}
\title{Create Interaction Nodes}
\usage{
create_interaction_nodes(
  interaction_variable,
  interaction_variable_type,
  interacting_variables,
  PMML
)
}
\arguments{
\item{interaction_variable}{list of interaction_variable names}

\item{interaction_variable_type}{list of interaction_variable types}

\item{interacting_variables}{list of interacting variable names}

\item{PMML}{the pmml that is appended to}
}
\value{
returns PMML with attached interaction nodes
}
\description{
Creates interaction nodes in the TransformationDictionary
}
