% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-model-export-to-pmml.R
\name{create_list_from_columns}
\alias{create_list_from_columns}
\title{Create list for each column of data}
\usage{
create_list_from_columns(data)
}
\arguments{
\item{data}{data.frame to be converted to list}
}
\value{
list with names set to column names from data
}
\description{
Utility function for creating list with elements for each column of data
}
