% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-model-export-to-pmml.R
\name{create_rcs_nodes}
\alias{create_rcs_nodes}
\title{Create RCS nodes}
\usage{
create_rcs_nodes(variable, rcs_variables, knots, PMML)
}
\arguments{
\item{variable}{list of original variables}

\item{rcs_variables}{list containing nested lists of rcs_variables for each original variable}

\item{knots}{list containing the knots}

\item{PMML}{the pmml that is appended to}
}
\value{
returns PMML with attached rcs nodes
}
\description{
Creates RCS nodes in the TransformationDictionary
}
