% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-to-pmml-test-utils.R
\name{csv_stringify_data_frame}
\alias{csv_stringify_data_frame}
\title{Converts a data.frame to a CSV string}
\usage{
csv_stringify_data_frame(data_frame)
}
\arguments{
\item{data_frame}{data.frame The data frame to convert}
}
\value{
string The CSV string
}
\description{
Converts a data.frame to a CSV string
}
