% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmml.R
\name{get_all_start_vars}
\alias{get_all_start_vars}
\title{Returns all the variable names, including the ones from children variables
that a variable depends on}
\usage{
get_all_start_vars(var_name, db_name, variable_details_sheet)
}
\arguments{
\item{variable_details_sheet}{data.frame A data frame that contains an entire
variables details sheet}

\item{derived_var}{string The name of the variable}
}
\value{
vector of strings Contains the names of the derived from variables
}
\description{
Returns all the variable names, including the ones from children variables
that a variable depends on
}
