% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{get_data_type}
\alias{get_data_type}
\title{Returns the PMML dataType attribute to use for a value}
\usage{
get_data_type(value)
}
\arguments{
\item{value}{The variable whose value we need the dataType attribute for}
}
\value{
string The valye of the dataType attribute
}
\description{
Returns the PMML dataType attribute to use for a value
}
