% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-derived-from-vars.R
\name{get_derived_from_vars}
\alias{get_derived_from_vars}
\title{Returns the variables needed to calculate a derived variable}
\usage{
get_derived_from_vars(derived_var, variable_details_sheet)
}
\arguments{
\item{derived_var}{string The name of the derived variable}

\item{variable_details_sheet}{data.frame A data frame containing a
variable details sheet}
}
\value{
list of named lists. Each named list represents one variable. The
named list has the following members:
1. name: string. The name of the variable
2. type: string. The type of the variable. Can have one of the following
values:
 * variable: Represents the name of a variable
 * constant: Represents a constant for example a string
 * table: Represents the name of a reference table
}
\description{
Returns the variables needed to calculate a derived variable
}
