% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-to-pmml-switch.R
\name{get_in_between_nodes}
\alias{get_in_between_nodes}
\title{Returns the nodes between a start and end node. 
Determining the in between nodes is done by looking at the order within a 
data frames of nodes, not by looking at their IDs.}
\usage{
get_in_between_nodes(tokens, start, end)
}
\arguments{
\item{tokens}{data.frame of nodes Used to determine the order}

\item{start}{data.frame containing the start node}

\item{end}{data.frame containing the end node}
}
\value{
A data.frame containing the in between nodes
}
\description{
Returns the nodes between a start and end node. 
Determining the in between nodes is done by looking at the order within a 
data frames of nodes, not by looking at their IDs.
}
