% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmml.R
\name{is_derived_var}
\alias{is_derived_var}
\title{Whether this variable details row is for a derived variable}
\usage{
is_derived_var(variable_details_row)
}
\arguments{
\item{variable_details_row}{data.frame One row from a variable details sheet}
}
\value{
boolean True if it is a derived variable, false otherwise
}
\description{
Whether this variable details row is for a derived variable
}
