% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeValues.R
\name{makeValues}
\alias{makeValues}
\title{Create Values element, most likely to add to a DataDictionary element}
\usage{
makeValues(value = NULL, displayValue = NULL, property = NULL,
  namespace = "4_2")
}
\arguments{
\item{value}{The 'value' attribute of each 'Value' element to be created in order.}

\item{displayValue}{The 'displayValue' attribute of each 'Value' element to be created in order.}

\item{property}{The 'property' attribute of each 'Value' element to be created in order.}

\item{namespace}{The namespace of the PMML model}
}
\value{
PMML Values elements.
}
\description{
Create Values element, most likely to add to a DataDictionary element
}
\details{
The 'makeValues' function is used the same way as the 'makeIntervals' function. If certain attributes for an
 element should not be included, they should be input in the list as NULL.
}
\examples{
# define 3 values, none with a 'displayValue' attribute and 1 value
# defined as 'invalid'. The 2nd one is 'valid' by default.
mv <- makeValues(list(1.1,2.2,3.3),list(NULL,NULL,NULL),
                 list("valid",NULL,"invalid"))
}
\author{
Tridivesh Jena
}
\seealso{
\code{\link{makeIntervals}} to make Interval child elements, \code{\link{addDFChildren}}
to add these xml fragments to the DataDictionary PMML element.
}

