% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-to-pmml-test-utils.R
\name{test_utils_run_generate_pmml_test}
\alias{test_utils_run_generate_pmml_test}
\title{Run a test for the get_pmml_string_from_r_file function}
\usage{
test_utils_run_generate_pmml_test(
  code,
  expected_pmml = NULL,
  expected_error = NULL,
  files = list()
)
}
\arguments{
\item{code}{string Contains the code for the R file that will be passed
into the function}

\item{expected_pmml}{string | NULL When testing if the function returns the
right PMML, this parameter should be set to the expected PMML}

\item{expected_error}{string | NULL When testing if the function throws the 
right error, this parameter should be set to the expected error}

\item{files}{named_list optional Use this parameter to set any additional
files that the test may need. For example, other R files that the main file
sources or CSV files that the main file reads in. The schema is below:
[file_name]: list(
    type: **R** for R files or **CSV** for csv files
    contents: string for R files or a data.frame for csv files
)}
}
\description{
Run a test for the get_pmml_string_from_r_file function
}
