% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-to-pmml-token.R
\name{token_is_na}
\alias{token_is_na}
\title{Checks whether the first token is NA}
\usage{
token_is_na(token)
}
\arguments{
\item{token}{}
}
\value{
boolean
}
\description{
It does this by checking whether the id column in the first token is NA
since the id column should normally always be defined
}
