% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-to-pmml-switch.R
\name{tokens_to_str}
\alias{tokens_to_str}
\title{Converts a data.frame of nodes into their corresponding code}
\usage{
tokens_to_str(parse_data)
}
\arguments{
\item{parse_data}{The data.frame of nodes to convert}

\item{A}{string containing the converted code}
}
\description{
Converts a data.frame of nodes into their corresponding code
}
