% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-model-export-to-pmml.R
\name{verify_length}
\alias{verify_length}
\title{Verify Lenght}
\usage{
verify_length(items)
}
\arguments{
\item{items}{vector of objects to verify the length of}
}
\value{
boolean representing if all items share identical length
}
\description{
Generic function to verify identical length of all passed items
}
